vocabulary-importer/vocabulary-importer-french/requirements.txt
streamlit==1.32.0
anthropic==0.19.1
pandas==2.2.0
pyperclip==1.8.2
python-dotenv==1.0.1
pydantic==2.6.1
streamlit-extras==0.3.6